# api/openapi.yaml
openapi: "3.0.2"
info:
  title: "Dogs app"
  version: "1.0"
  description: "This is the API definition of the dogs owners app."
servers:
  - url: https://api.server.test/v1
paths:
...
  /users/{userId}:
    put:
      tags:
        - "Users"
      description: A user can update only his own profile.
      summary: "Update an existing user"
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int64
          required: true
          description: Numeric ID of the user to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        description: "User object that needs to be updated"
      responses:
        204:
          description: The user was susscesfully updated
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
    get:
      tags:
        - "Users"
      summary: "Return user by userId"
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int64
          required: true
          description: Numeric ID of the user to get
      responses:
        200:
          description: A user with id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: "#/components/responses/401"
        404:
          $ref: "#/components/responses/404"
    delete:
      tags:
        - "Users"
      description: A user can delete his own profile.
      summary: "Delete user with userId"
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int64
          required: true
          description: Numeric ID of the user to delete
      responses:
        204:
          description: A user with id was deleted
        400:
          $ref: '#/components/responses/400'
        403:
          $ref: '#/components/responses/403'
        401:
          $ref: "#/components/responses/401"
        404:
          $ref: "#/components/responses/404"
...
  /path/{pathId}/reviews:
    get:
      tags:
        - "Paths"
      summary: "Return all review of walking path by pathId"
      parameters:
        - in: path
          name: pathId
          schema:
            type: integer
            format: int64
          required: true
          description: Numeric ID of the path
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
          description: A list of reviews for a path with pathId
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: "#/components/responses/401"
        404:
          $ref: "#/components/responses/404"
    post:
      tags:
        - "Paths"
      description: A user can add review to the walking path.
      summary: "Add review to path with pathId"
      parameters:
        - in: path
          name: pathId
          schema:
            type: integer
            format: int64
          required: true
          description: Numeric ID of the path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Review"
        description: "Review object that needs to be created"
      responses:
        201:
          description: The review was susscesfully added
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
components:
  responses:
    401:
      description: User unauthenticated
    403:
      description: User not permitted
    404:
      description: Resource not found
    400:
      description: Invalid Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        name:
          type: string
          readOnly: true
        gender:
          $ref: '#/components/schemas/Gender'
        dateOfBirth:
          type: string
          format: date
        bio:
          type: string
        status:
          readOnly: true
          $ref: '#/components/schemas/OnlineStatus'
        profilePicture:
          type: string
          readOnly: true
        latestLocation:
          $ref: '#/components/schemas/Coordinate'
          readOnly: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        conversations:
          readOnly: true
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
        dogs:
          readOnly: true
          type: array
          items:
            $ref: '#/components/schemas/Dog'
    UserOverview:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        name:
          type: string
          readOnly: true
        profilePicture:
          type: string
          readOnly: true
        status:
          readOnly: true
          $ref: '#/components/schemas/OnlineStatus'
    Dog:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        breed:
          type: string
        sex:
          $ref: '#/components/schemas/Gender'
        dateOfBirth:
          type: string
          format: date
        weight:
          type: number
          format: double
        description:
          type: string
        color:
          type: string
      required:
        - name
        - breed
        - sex
...
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        creator:
          $ref: '#/components/schemas/User'
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        text:
          type: string
        timeStamp:
          type: string
          format: date-time
          readOnly: true
      required:
        - rating
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        sender:
          $ref: '#/components/schemas/UserOverview'
        receiver:
          $ref: '#/components/schemas/UserOverview'
        message:
          type: string
        timeStamp:
          type: string
          format: date-time
          readOnly: true
        unread:
          type: boolean
          readOnly: true
      required:
        - message
        - receiver
        - sender
    Conversation:
      type: object
      properties:
        participant:
          $ref: '#/components/schemas/UserOverview'
        lastMessage:
          $ref: '#/components/schemas/ChatMessage'
...
    ErrorResponse:
      type: object
      properties:
        reason:
          type: string
        statusCode:
          type: number
    Gender:
      type: string
      enum: [MALE, FEMALE, OTHER]
    OnlineStatus:
      type: string
      enum: [ ONLINE, OFFLINE ]
    Tag:
      type: object
      properties:
        name:
          type: string
        tagType:
          type: string
          enum:
            - OFFERING
            - LOOKINGFOR
    Coordinate:
      type: object
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
      required:
        - latitude
        - longitude
...